Export helper which transforms platform specific hotkey conversions
On Windows and macOS, other keys are considered as primary actions. For example, the "command" key and the "Ctrl" key on Windows. When using "primary" as key instead, the hotkey will automatically detect the host OS and choose the matching primary key.
The mapping for "secondary" is debatable, as macOS "option" is hard to handle as it creates special characters.